Refractor: Infer Generic Type arguments
I will start removing a bunch of warnings in the next few commits as I belive this helps preparing for the 1.16 update. This means a lot of refractors are going to happen and the code will not be compatible with what it was. In case this should backfire, it can be easily reverted.